backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn

# File handling
python-multipart==0.0.9
aiofiles==23.2.1

# Config
python-dotenv==1.0.1

# Google Sheets
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0

# Document generation (Teskom)
docxtpl
python-docx
Pillow

# Utils
jinja2
